backend/requirements.txt
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
snowflake-connector-python>=3.6
google-generativeai>=0.7
